Principal: Manejo de persistencia de usuario (login user: demo pass: demo123)

+++ app/src/main/java/com/cineplusapp/cineplusspaapp/data/model/Product.kt
@@ -7,6 +7,6 @@ data class Product(
     val name: String,
     val description: String = "",
     val imageUrl: String? = null,
-    val price: Double = 0.0,
+    val price: Int = 0,
     val type: ProductType = ProductType.COMBO
 )
